requirements.txt
flask==3.0.2
google-generativeai==0.3.2
pandas==2.2.1
numpy==1.26.4 